pos-backend/requirements.txt
pydantic-settings
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
pydantic
redis
celery
python-dotenv
requests
cachetools
